requirements.txt
uvicorn
keras
matplotlib
numpy
fastapi[all]
pillow
tensorflow
